+++ docs/changelog.md
@@ -28,6 +28,7 @@
 - Use correct table name when exporting to PostgreSQL
 - Carefully close all fobjs in pgimport/pgexport
 - Added CSV dialect "excel-semicolon"
+- Improved PostgreSQL import from CSV (pgimport) when dealing with null values
 
 ### Command-Line Interface